mv map_3b_kern to mgp/utils.py. TODO: write numba version for 2b & energy block
